+++ src/test/java/com/paypal/heapdumptool/sanitizer/HeapDumpSanitizerSyntheticTest.java
@@ -1,5 +1,6 @@
 package com.paypal.heapdumptool.sanitizer;
 
+import com.paypal.heapdumptool.utils.DataSize;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import picocli.CommandLine;